457: move all cache related code to src/cache
